cleanup huffman code
